@@ -1,7 +1,7 @@
 //= require_self
 //= require jquery
 //= require jquery_ujs
-//= require jquery.ui.all
+//= require jquery-ui
 //= require jquery-fileupload
 //= require jquery.onecarousel.min
 //= require bootstrap
