@@ -1,12 +1,15 @@
-class LinkCell < Cell::Rails
-  helper LinksHelper
-  # cache :item, :expires_in => 6.hours do |cell, item|
-  #   "#{item[:link].id}-#{item[:link].updated_at}-xx"
-  # end
+# class LinkCell < Cell::ViewModel
+#   include LinksHelper
 
-  def item(args)
-    @link = args[:link]
-    render
-  end
+#   property :link
 
-end
+#   def show
+#     render
+#   end
+
+#   # def item(args)
+#   #   @link = args[:link]
+#   #   render
+#   # end
+
+# end
