@@ -7,12 +7,16 @@ class ActsAsVotableMigration < ActiveRecord::Migration
 
       t.boolean :vote_flag
       t.string :vote_scope
+      t.integer :vote_weight
 
       t.timestamps
     end
 
-    add_index :votes, [:votable_id, :votable_type]
-    add_index :votes, [:voter_id, :voter_type]
+    if ActiveRecord::VERSION::MAJOR < 4
+      add_index :votes, [:votable_id, :votable_type]
+      add_index :votes, [:voter_id, :voter_type]
+    end
+
     add_index :votes, [:voter_id, :voter_type, :vote_scope]
     add_index :votes, [:votable_id, :votable_type, :vote_scope]
   end
